+++ src/Resource/Collection/ArrayResourceCollection.php
@@ -11,13 +11,13 @@
 
 namespace Puli\Repository\Resource\Collection;
 
-use Assert\Assertion;
 use InvalidArgumentException;
 use IteratorAggregate;
 use OutOfBoundsException;
 use Puli\Repository\Api\Resource\Resource;
 use Puli\Repository\Api\ResourceCollection;
 use Puli\Repository\Api\UnsupportedResourceException;
+use Puli\Repository\Assert\Assert;
 use Puli\Repository\Resource\Iterator\ResourceCollectionIterator;
 
 /**
@@ -119,8 +119,7 @@ class ArrayResourceCollection implements IteratorAggregate, ResourceCollection
      */
     public function replace($resources)
     {
-        Assertion::isTraversable($resources);
-        Assertion::allIsInstanceOf($resources, 'Puli\Repository\Api\Resource\Resource');
+        Assert::allIsInstanceOf($resources, 'Puli\Repository\Api\Resource\Resource');
 
         $this->resources = is_array($resources) ? $resources : iterator_to_array($resources);
     }
@@ -130,8 +129,7 @@ class ArrayResourceCollection implements IteratorAggregate, ResourceCollection
      */
     public function merge($resources)
     {
-        Assertion::isTraversable($resources);
-        Assertion::allIsInstanceOf($resources, 'Puli\Repository\Api\Resource\Resource');
+        Assert::allIsInstanceOf($resources, 'Puli\Repository\Api\Resource\Resource');
 
         // only start merging after validating all resources
         foreach ($resources as $resource) {