+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 * 1.0.0-next (@release_date@)
 
  * added `NullRepository`
+ * removed dependency to beberlei/assert
 
 * 1.0.0-beta (2015-01-12)